Req 18
Every time that an application changes its status, the system must send
a message to both the manager and the explorer involved to let them
know of the change. The message must be stored in folder “notification
folder”.

MessageService sendNotification() was fixed!
— Test were also fixed in MessageServiceTest, regarding this last change
TripApplicationService
changeApplicationStatus() was fixed!
PopulateXML was added a folder to explorer1