Req 12.1 Finally!
StageManagerController was added
TagValueManagerController was added
TripManagerController was fixed: now it provides the actorWS and does
not automatically cancels every trip that its changed

Now there is an specific query in TagRepository to get the Tags a Trip
uses

StageServices was fixed. Now it properly computes the number. When
editing a stage, the price of the trip must be updated (fixed)

TagValueService was fixed. Create was change and now it only requires
one param. Save was missing.

Association with TagValue was incorrect and not mappedBy!