Part 2 of Fixing Corchu's Bugs Series 🐞
- Now a trip may be tagged using repeating tags

- Now, it is checked that a trip may be only edited by its
corresponding manager and that only for trips that aren’t over.

- Added restriction notes to TripService’s save() method to ensure that
the publication date is before the starting date and that the ending
date is after the starting date.

- Now button save is enabled only when all required fields are
completed in actor/edit.jsp (for all the cases, not just an
unregistered actor)

- audit/edit.jsp was removed the possibility of changing from one trip
to another, so it does not show a strange error

- Added the restriction that only one audit may be created by an
Auditor for the same trip

NEW: Trip now features price calculation with Taxation! (review this)